T60: fix touchpad option
Code used 'int' as return type, but the cmos option is only one
bit. get_option returned with the value in bit 0-7, but all remaining
bits were left unitialized by get_option(). fix this by using char
as type.

Reviewed-on: http://review.coreboot.org/5
Tested-by: build bot (Jenkins)

+++ src/mainboard/lenovo/t60/mainboard.c
@@ -39,7 +39,7 @@
 static void mainboard_enable(device_t dev)
 {
 	device_t dev0, idedev;
-	int touchpad;
+	u8 touchpad;
 
 	/* enable Audio */
 	h8_set_audio_mute(0);
